Change build-swift-cmake.py to use the system debugserver by default.
This also changes the flag to be --no-system-debugserver if we
don't want to use the system debugserver.

It also prevents the script from fighting with the "use the system
debugserver" and "include *no* debugserver at all" option.  The
no debugserver option wins in this case now.

rdar://problem/23630932

apple-llvm-split-commit: 1594ac06c1a0043c6eb05648a300c2cce4fb814d
apple-llvm-split-dir: lldb/