Fix the lldbinline tests so they make well-formed Makefiles.
lldbinline tests previously did not run correctly unless there was already a
Makefile for them.  This was because the syntax of the emitted Makefile made the
default make rule be the "cleanup" rule, which is pretty unhelpful.  Now the
default rule is the one included from Makefile.rules, which is much better.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@258763 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 5c036ccef6c0aba7e0e21ea3d0cb862165e9b23d
apple-llvm-split-dir: lldb/