Lex: Return "" when HeaderMap::lookupFilename fails
Change getString() to return Optional<StringRef>, and change
lookupFilename() to return an empty string if either one of the prefix
and suffix can't be found.

This is a more robust follow-up to r261461, but it's still not entirely
satisfactory.  Ideally we'd report that the header map is corrupt;
perhaps something for a follow-up.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@261596 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 6d0c65dab43962dff862d8263333feb385d71c96)

apple-llvm-split-commit: 04e3d9442e18aa5d5c1d0587672022e70d7a2e29
apple-llvm-split-dir: clang/