[analyzer] Improve pointer arithmetic checker.
This patch is intended to improve pointer arithmetic checker.
From now on it only warns when the pointer arithmetic is likely to cause an
error. For example when the pointer points to a single object, or an array of
derived types.

Differential Revision: http://reviews.llvm.org/D14203

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@261632 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit e17e2519c1697b449e3e2c7b4746bc0f786505ca)

apple-llvm-split-commit: 444987001880f5826edee2c8a7d10ac0717c8397
apple-llvm-split-dir: clang/