Remove an unnecessary workaround introduced in r259975. (NFC)
Now that LLVM r259973 allows replacing a temporary type with another
temporary we can rely on the original implementation.

It is possible for enums to be created as part of
their own declcontext. In this case a FwdDecl will be created
twice. This doesn't cause a problem because both FwdDecls are
entered into the ReplaceMap: finalize() will replace the first
FwdDecl with the second and then replace the second with
complete type.

Thanks to echristo for pointing this out.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@261657 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit c473d6dc331e6e199803d742e3bdd28035191fd0)

# Conflicts:
#	lib/CodeGen/CGDebugInfo.cpp

apple-llvm-split-commit: f953dddad01770bb89f9fadb809ed6842f6aa5ba
apple-llvm-split-dir: clang/