adjust Xcode build to avoid rebuilding solely due to a logical no-op revision file rewrite.
The github version of LLDB has a versioning scheme more in line with
Swift's versioning.  The Xcode build was not being particulary smart
about it, overwriting the version file each time. This happened whether
the content changed or not.  This would cause a file timestamp check to
force the LLDB driver to be rebuilt even if nothing changed, as the
version stamp file was getting rewritten.

This change now writes the content to a temporary file, and only
rewrites the original file if (a) the contents changed or (b) there
is no original file.  This prevents Xcode from wanting to rebuild
the LLDB driver if nothing else changed.

apple-llvm-split-commit: dc77c6556e493d940671a7f2018095741b46d070
apple-llvm-split-dir: lldb/