Fix destructor definition of invalid classes
The declaration of the destructor of an invalid class was not properly marked
as noexcept. As a result, the definition of the same destructor, which was
properly implicitly marked as noexcept, would not match the definition.
This would cause the definition CXXDestructorDecl to be matked as invalid
and omited from the AST.

Differential Revision: http://reviews.llvm.org/D17988

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@263639 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit dc4b244bbdddbabdcdc66306a5fbb12470cbf4d8)

apple-llvm-split-commit: fd9eb456ae71faa9638edcac94a29bc5bca52d6b
apple-llvm-split-dir: clang/