[Objective-c] Fix a crash in WeakObjectProfileTy::getBaseInfo.
The crash occurs in WeakObjectProfileTy::getBaseInfo when getBase() is
called on an ObjCPropertyRefExpr object whose receiver is an interface.
This commit fixes the crash by checking the type of the receiver and
setting IsExact to true if it is an interface.

rdar://problem/25208167

Differential Revision: http://reviews.llvm.org/D18268

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@263818 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 1d22cda9dbf116c761db465ee2c7ce50acb7863d)

apple-llvm-split-commit: cce9fe9fe6eaf233ba0049a65f6b1936d7538b57
apple-llvm-split-dir: clang/