[LLE] Check for mismatching types between the store and the load earlier
isDependenceDistanceOfOne asserts that the store and the load access
through the same type.  This function is also used by
removeDependencesFromMultipleStores so we need to make sure we filter
out mismatching types before reaching this point.

Now we do this when the initial candidates are gathered.

This is a refinement of the fix made in r262267.

Fixes PR27048.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@264313 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 1123cc74dff602ee5f52a1542433e18a5416dc55)

apple-llvm-split-commit: 10def6b238e9c6a20fa88b8352a14b2e555e9f33
apple-llvm-split-dir: llvm/